dncnn_image_compression_streamlit-main/requirements.txt
absl-py
altair
aniso8601
annotated-types
anyio
argon2-cffi
argon2-cffi-bindings
arrow
asttokens
astunparse
async-lru
attrs
Babel
bcrypt
beautifulsoup4
bidict
bleach
blinker
cachetools
certifi
cffi
charset-normalizer
click
colorama
comm
contourpy
cycler
debugpy
decorator
defusedxml
executing
fastjsonschema
filelock
Flask
Flask-Cors
Flask-RESTful
Flask-SocketIO
flatbuffers
fonttools
fqdn
fsspec
gast
gitdb
GitPython
google-pasta
grpcio
h11
h5py
htmlmin
idna
ImageHash
imageio
ipykernel
ipython
ipywidgets
isoduration
itsdangerous
jedi
Jinja2
joblib
json5
jsonpointer
jsonschema
jsonschema-specifications
jupyter
jupyter-console
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab-widgets
jupyterlab_pygments
jupyterlab_server
keras
kiwisolver
lazy_loader
libclang
Markdown
markdown-it-py
MarkupSafe
matplotlib
matplotlib-inline
mdurl
memory-profiler
missingno
mistune
ml-dtypes
mpmath
multimethod
namex
narwhals
nbclient
nbconvert
nbformat
nest-asyncio
networkx
notebook
notebook_shim
numpy
opencv-contrib-python
opencv-python
opencv-python-headless
opt_einsum
optree
overrides
packaging
pandas
pandocfilters
parso
phik
pillow
platformdirs
prometheus-client
prompt-toolkit
protobuf
psutil
pure-eval
pyarrow
pycparser
pydantic
pydantic-settings
pydantic_core
pydeck
pygame
Pygments
pyparsing
pytesseract
python-dateutil
python-dotenv
python-engineio
python-json-logger
python-socketio
pytz
PyWavelets
pywinpty
PyYAML
pyzmq
qtconsole
QtPy
referencing
requests
rfc3339-validator
rfc3986-validator
rich
rpds-py
scapy
scikit-fuzzy
scikit-image
scikit-learn
scipy
seaborn
Send2Trash
setuptools
simple-websocket
six
smmap
sniffio
soupsieve
stack-data
streamlit
streamlit-option-menu
sympy
tangled-up-in-unicode
tenacity
tensorboard
tensorboard-data-server
tensorflow
tensorflow_intel
termcolor
terminado
tesseract
threadpoolctl
tifffile
tinycss2
toml
torch
torchaudio
torchvision
tornado
tqdm
traitlets
types-python-dateutil
typing_extensions
tzdata
uri-template
urllib3
visions
watchdog
wcwidth
webcolors
webencodings
websocket-client
Werkzeug
wheel
widgetsnbextension
wrapt
wsproto
